pyproject.toml
[project]
name = "falcon-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-ckeditor-5>=0.2.18",
    "django-jazzmin>=3.0.1",
    "django-jsonform>=2.23.2",
    "django-rosetta>=0.10.2",
    "factory-boy>=3.3.3",
    "faker>=37.11.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
